pyproject.toml
[tool.poetry]
name = "dart_object_extension"
version = "0.0.1"
description = "dart_object_extension Project."
authors = ["JosephNK <nkw0608@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "scripts"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
ruamel-yaml = "^0.19.1"
gitpython = "^3.1.0"

[tool.poetry.scripts]
sync-version = "scripts.sync_version:main"
publish = "scripts.publish:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"